backend/requirements.txt
Flask
mysql-connector-python
twilio
face-recognition
opencv-python
PyQt5
